\name{demo1}
\alias{demo1}
\title{Demo for XDATA inference in CharityNet via graph embedding.}
\usage{
  demo1(method = 1)
}
\arguments{
  \item{method}{1: methd1, 2: method2}
}
\description{
  Demo for XDATA inference in CharityNet via graph
  embedding.
}
\seealso{
  \code{\link{demo2}} for XDATA Time Series of Graphs
  anomaly detection in CharityNet via scan statistics.
}

