\name{demo2}
\alias{demo2}
\title{Demo for XDATA Time Series of Graphs anomaly detection in CharityNet via
scan statistics.}
\usage{
  demo2(plotmap = FALSE)
}
\arguments{
  \item{plotmap}{flag to draw map or not.}
}
\description{
  Demo for XDATA Time Series of Graphs anomaly detection in
  CharityNet via scan statistics.
}
\seealso{
  \code{\link{demo1}} for XDATA inference in CharityNet via
  graph embedding.
}

