\name{getTrans.at.T}
\alias{getTrans.at.T}
\title{Get subset of the list of transactions that takes place from begin to end}
\usage{
  getTrans.at.T(trans.list, begin, end)
}
\arguments{
  \item{trans.list}{list of all transactions}

  \item{begin}{beginning of the time interval}

  \item{end}{end of the time interval}
}
\value{
  subset of the list of transactions
}
\description{
  Get subset of the list of transactions that takes place
  from begin to end
}

