\docType{package}
\name{jhuxdata-package}
\alias{jhuxdata}
\alias{jhuxdata-package}
\title{The jhuxdata Package}
\description{
  XDATA JHU demo package.
}
\details{
  \tabular{ll}{ Package: \tab jhuxdata\cr Type: \tab
  Package\cr Version: \tab 0.2\cr Date: \tab 2013-06-13\cr
  Depends: \tab igraph, reshape2, animation, gridExtra,
  ggplot2, MCMCpack, mclust, Matrix, irlba, plyr, maps,
  R.utils, cluster, mapproj, R.utils, chron\cr License:
  \tab GPL (>= 3)\cr LazyLoad: \tab yes\cr }

  See 'helloworld' for a quick demo, and 'demo1' & 'demo2'
  for step by step details.
}
\author{
  Youngser Park \email{youngser@jhu.edu}
}
\keyword{package}

