\name{normalize_verts.at.T.with_tau}
\alias{normalize_verts.at.T.with_tau}
\title{stats_at_tau_window the scan stats
as a list in tau length time window (t-tau to t)}
\usage{
  normalize_verts.at.T.with_tau(stats_at.T, mean_phi_tau,
    sd_phi_tau)
}
\description{
  stats_at_tau_window the scan stats as a list in tau
  length time window (t-tau to t)
}

