\name{plotmap}
\alias{plotmap}
\title{Function to draw US maps with CharityNet transactions}
\usage{
  plotmap(dat, donor, charity, mcout = NULL, rng = NULL,
    main = NULL)
}
\arguments{
  \item{dat}{data frame of transactions}

  \item{donor}{data frame of donors}

  \item{charity}{data frame of charities}

  \item{mcout}{vector of labels for coloring the mapping
  object}

  \item{rng}{vector of min and max range of transaction
  amount}

  \item{main}{character for the plot title}
}
\description{
  Function to draw US maps with CharityNet transactions
}

